Check if scripts are running inside the Jenkins env

+++ bin/Docker_CleanUp.sh
@@ -1,8 +1,11 @@
 #!/bin/bash
 
-# Obtain all containers spawned from the latest version of our image
-for label in $(docker ps --filter "ancestor=${DOCKER_IMAGE_LATEST}" --format "{{ .Labels }}") ; do
-  swarm=$(echo ${label} | tr ',' "\n" | grep 'com.docker.stack.namespace' | awk -F= '{ print $2; }')
-  echo "Found swarm [${swarm}]"
-  docker rm ${swarm}
-done
+# Check whether there is any leftover docker stack from last build
+if [ -n ${DOCKER_IMAGE_LATEST} ] then 
+	# Obtain all containers spawned from the latest version of our image
+	for label in $(docker ps --filter "ancestor=${DOCKER_IMAGE_LATEST}" --format "{{ .Labels }}") ; do
+	  swarm=$(echo ${label} | tr ',' "\n" | grep 'com.docker.stack.namespace' | awk -F= '{ print $2; }')
+	  echo "Found swarm [${swarm}]"
+	  docker rm ${swarm}
+	done
+fi
